LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbcursor_continue_objectstore4-expected.txt


PASS IDBCursor.continue() - object store - attempt to iterate to the next record when the direction is set for the previous record
